fix: json container default values
