+++ web/crux-ui/src/models/container.ts
@@ -1,4 +1,3 @@
-import { v4 as uuid } from 'uuid'
 import { imageName } from './registry'
 
 export const CONTAINER_STATE_VALUES = ['running', 'waiting', 'exited', 'removed'] as const
@@ -74,8 +73,6 @@ export type ContainerPortRange = {
   internal: PortRange
   external: PortRange
 }
-
-export type JsonKeyValue = { [key: string]: string }
 
 export const CONTAINER_NETWORK_MODE_VALUES = ['none', 'host', 'bridge'] as const
 export type ContainerNetworkMode = (typeof CONTAINER_NETWORK_MODE_VALUES)[number]
@@ -316,75 +313,6 @@ export const filterContains = (
 export const filterEmpty = (filterValues: string[], filters: ContainerConfigKey[]): boolean =>
   filterValues.filter(x => filters.includes(x as ContainerConfigKey)).length > 0
 
-export type JsonInitContainer = {
-  name: string
-  image: string
-  command?: string[]
-  args?: string[]
-  environment?: JsonKeyValue
-  useParentConfig?: boolean
-  volumes?: JsonInitContainerVolumeLink[]
-}
-
-export type JsonContainerConfigLog = {
-  driver: ContainerLogDriverType
-  options: JsonKeyValue
-}
-
-export type JsonMarker = {
-  service: JsonKeyValue
-  deployment: JsonKeyValue
-  ingress: JsonKeyValue
-}
-
-export type JsonInitContainerVolumeLink = Omit<InitContainerVolumeLink, 'id'>
-export type JsonContainerConfigPortRange = Omit<ContainerPortRange, 'id'>
-export type JsonContainerConfigPort = Omit<Port, 'id'>
-export type JsonContainerConfigVolume = Omit<Volume, 'id'>
-export type JsonContainerConfigSecretKey = Omit<UniqueSecretKey, 'id'>
-
-export type JsonContainerConfig = {
-  // common
-  name?: string
-  environment?: JsonKeyValue
-  secrets?: JsonContainerConfigSecretKey[]
-  routing?: ContainerConfigRouting
-  expose?: ContainerConfigExposeStrategy
-  user?: number
-  workingDirectory?: string
-  tty?: boolean
-  configContainer?: ContainerConfigContainer
-  ports?: JsonContainerConfigPort[]
-  portRanges?: JsonContainerConfigPortRange[]
-  volumes?: JsonContainerConfigVolume[]
-  commands?: string[]
-  args?: string[]
-  initContainers?: JsonInitContainer[]
-  capabilities?: JsonKeyValue
-  storage?: ContainerStorage
-  expectedState?: ExpectedContainerState
-
-  // dagent
-  logConfig?: JsonContainerConfigLog
-  restartPolicy?: ContainerRestartPolicyType
-  networkMode?: ContainerNetworkMode
-  networks?: string[]
-  dockerLabels?: JsonKeyValue
-
-  // crane
-  deploymentStrategy?: ContainerDeploymentStrategyType
-  customHeaders?: string[]
-  proxyHeaders?: boolean
-  useLoadBalancer?: boolean
-  extraLBAnnotations?: JsonKeyValue
-  healthCheckConfig?: ContainerConfigHealthCheck
-  resourceConfig?: ResourceConfig
-  annotations?: JsonMarker
-  labels?: JsonMarker
-}
-
-export type ConcreteJsonContainerConfig = Omit<JsonContainerConfig, 'secrets'>
-
 export const stringResettable = (base: string, concrete: string): boolean => {
   if (!concrete) {
     return false
@@ -419,323 +347,6 @@ export const booleanResettable = (base: boolean, concrete: boolean): boolean => 
   }
 
   return base !== concrete
-}
-
-const keyValueArrayToJson = (list: UniqueKeyValue[]): JsonKeyValue =>
-  list?.reduce((prev, it) => ({ ...prev, [it.key]: it.value }), {})
-
-const keyArrayToJson = (list: UniqueKey[]): string[] => list?.map(it => it.key)
-
-const removeId = <T extends { id: string }>(item: T): Omit<T, 'id'> => {
-  const newItem: T = { ...item }
-  delete newItem.id
-
-  return newItem
-}
-
-export const containerConfigToJsonConfig = (config: ContainerConfigData): JsonContainerConfig => ({
-  // common
-  name: config.name,
-  environment: keyValueArrayToJson(config.environment),
-  // secrets are ommited
-  routing: config.routing,
-  expose: config.expose,
-  user: config.user,
-  workingDirectory: config.workingDirectory,
-  tty: config.tty,
-  configContainer: config.configContainer,
-  ports: config.ports?.map(it => removeId(it)),
-  portRanges: config.portRanges?.map(it => removeId(it)),
-  volumes: config.volumes?.map(it => removeId(it)),
-  commands: keyArrayToJson(config.commands),
-  args: keyArrayToJson(config.args),
-  initContainers: config.initContainers?.map(container => ({
-    ...removeId(container),
-    command: keyArrayToJson(container.command),
-    args: keyArrayToJson(container.args),
-    environment: keyValueArrayToJson(container.environment),
-    volumes: container.volumes?.map(vit => removeId(vit)),
-  })),
-  capabilities: keyValueArrayToJson(config.capabilities),
-  storage: config.storage,
-
-  // dagent
-  logConfig: config.logConfig
-    ? {
-        ...config.logConfig,
-        options: keyValueArrayToJson(config.logConfig?.options),
-      }
-    : null,
-  restartPolicy: config.restartPolicy,
-  networkMode: config.networkMode,
-  networks: keyArrayToJson(config.networks),
-  dockerLabels: keyValueArrayToJson(config.dockerLabels),
-  expectedState: config.expectedState,
-
-  // crane
-  deploymentStrategy: config.deploymentStrategy,
-  customHeaders: keyArrayToJson(config.customHeaders),
-  proxyHeaders: config.proxyHeaders,
-  useLoadBalancer: config.useLoadBalancer,
-  extraLBAnnotations: keyValueArrayToJson(config.extraLBAnnotations),
-  healthCheckConfig: config.healthCheckConfig,
-  resourceConfig: config.resourceConfig,
-  annotations: config.annotations
-    ? {
-        deployment: keyValueArrayToJson(config.annotations.deployment),
-        service: keyValueArrayToJson(config.annotations.service),
-        ingress: keyValueArrayToJson(config.annotations.ingress),
-      }
-    : null,
-  labels: config.labels
-    ? {
-        deployment: keyValueArrayToJson(config.labels.deployment),
-        service: keyValueArrayToJson(config.labels.service),
-        ingress: keyValueArrayToJson(config.labels.ingress),
-      }
-    : null,
-})
-
-export const concreteContainerConfigToJsonConfig = (config: ConcreteContainerConfig): ConcreteJsonContainerConfig => {
-  const json = containerConfigToJsonConfig(config)
-
-  delete json.secrets
-
-  return json
-}
-
-const mergeKeyValuesWithJson = (items: UniqueKeyValue[], json: JsonKeyValue): UniqueKeyValue[] => {
-  items = items ?? []
-
-  if (!json || Object.keys(json).length < 1) {
-    return []
-  }
-
-  let modified = false
-  const result = []
-  const jsonKeys = Object.keys(json)
-
-  jsonKeys.forEach(key => {
-    const value = json[key]
-
-    const byKey = items.find(it => it.key === key)
-    if (!byKey) {
-      const byValue = items.find(it => it.value === value)
-
-      result.push({
-        key,
-        value,
-        id: byValue?.id ?? uuid(),
-      })
-
-      modified = true
-    } else {
-      if (byKey.value !== value) {
-        modified = true
-      }
-
-      result.push({
-        key,
-        value,
-        id: byKey.id,
-      })
-    }
-  })
-
-  const removed = items.filter(it => !jsonKeys.includes(it.key))
-  if (removed.length > 0) {
-    modified = true
-  }
-
-  return modified ? result : items
-}
-
-const mergeKeysWithJson = (items: UniqueKey[], json: string[]): UniqueKey[] => {
-  items = items ?? []
-
-  if (!json || Object.entries(json).length < 1) {
-    return []
-  }
-
-  let modified = false
-  const result = []
-  json.forEach(entry => {
-    const byKey = items.find(it => it.key === entry)
-    if (!byKey) {
-      result.push({
-        key: entry,
-        id: uuid(),
-      })
-
-      modified = true
-    } else {
-      if (byKey.key !== entry) {
-        modified = true
-      }
-
-      result.push({
-        key: entry,
-        id: byKey.id,
-      })
-    }
-  })
-
-  const jsonKeys = Object.keys(json)
-  const removed = items.filter(it => !jsonKeys.includes(it.key))
-  if (removed.length > 0) {
-    modified = true
-  }
-
-  return modified ? result : items
-}
-
-const mergeSecretsWithJson = (secrets: UniqueSecretKey[], json: JsonContainerConfigSecretKey[]): UniqueSecretKey[] => {
-  secrets = secrets ?? []
-
-  json?.forEach(it => {
-    const index = secrets.findIndex(sec => sec.key === it.key)
-
-    if (index > -1) {
-      const sec = secrets[index]
-      secrets[index] = {
-        ...sec,
-        ...it,
-      }
-    } else {
-      secrets.push({
-        ...it,
-        id: uuid(),
-      })
-    }
-  })
-
-  return secrets
-}
-
-const mergeInitContainersWithJson = (containers: InitContainer[], json: JsonInitContainer[]): InitContainer[] => {
-  if (!json) {
-    return containers ?? []
-  }
-
-  containers = containers ?? []
-
-  json?.forEach(cont => {
-    const contIndex = containers.findIndex(it => it.name === cont.name)
-
-    if (contIndex > -1) {
-      const current = containers[contIndex]
-
-      containers[contIndex] = {
-        ...current,
-        args: mergeKeysWithJson(current.args, cont.args),
-        command: mergeKeysWithJson(current.command, cont.command),
-        environment: mergeKeyValuesWithJson(current.environment, cont.environment),
-        volumes: cont.volumes?.map(volume => {
-          const currentVol = current.volumes?.find(it => it.name === volume.name)
-
-          return {
-            ...volume,
-            id: currentVol?.id ?? uuid(),
-          }
-        }),
-      }
-    } else {
-      containers.push({
-        ...cont,
-        id: uuid(),
-        command: cont.command?.map(it => ({ id: uuid(), key: it })) ?? [],
-        args: cont.args?.map(it => ({ id: uuid(), key: it })) ?? [],
-        environment:
-          Object.keys(cont.environment ?? {}).map(it => ({
-            key: it,
-            value: cont.environment[it],
-            id: uuid(),
-          })) ?? [],
-        useParentConfig: cont.useParentConfig ?? false,
-        volumes: cont.volumes?.map(it => ({ ...it, id: uuid() })) ?? [],
-      })
-    }
-  })
-
-  return containers
-}
-
-export const mergeJsonConfigToConcreteContainerConfig = (
-  config: ConcreteContainerConfig,
-  json: ConcreteJsonContainerConfig,
-): ConcreteContainerConfig => {
-  const result: ConcreteContainerConfig = {
-    ...config,
-    ...json,
-    environment: mergeKeyValuesWithJson(config.environment, json.environment),
-    extraLBAnnotations: mergeKeyValuesWithJson(config.extraLBAnnotations, json.extraLBAnnotations),
-    capabilities: mergeKeyValuesWithJson(config.capabilities, json.capabilities),
-    commands: mergeKeysWithJson(config.commands, json.commands),
-    customHeaders: mergeKeysWithJson(config.customHeaders, json.customHeaders),
-    networks: mergeKeysWithJson(config.networks, json.networks),
-    args: mergeKeysWithJson(config.args, json.args),
-    logConfig: json.logConfig
-      ? {
-          ...config.logConfig,
-          ...json.logConfig,
-          options: mergeKeyValuesWithJson(config.logConfig?.options, json.logConfig?.options),
-        }
-      : null,
-    initContainers: mergeInitContainersWithJson(config.initContainers, json.initContainers),
-    ports: !json.ports
-      ? []
-      : json.ports.map(it => ({
-          ...it,
-          id: uuid(),
-        })),
-    portRanges: !json.portRanges
-      ? []
-      : json.portRanges.map(it => ({
-          ...it,
-          id: uuid(),
-        })),
-    dockerLabels: mergeKeyValuesWithJson(config.dockerLabels, json.dockerLabels),
-    labels: json.labels
-      ? {
-          deployment: mergeKeyValuesWithJson(config.labels?.deployment, json.labels.deployment),
-          service: mergeKeyValuesWithJson(config.labels?.service, json.labels.service),
-          ingress: mergeKeyValuesWithJson(config.labels?.ingress, json.labels.ingress),
-        }
-      : null,
-    annotations: json.annotations
-      ? {
-          deployment: mergeKeyValuesWithJson(config.annotations?.deployment, json.annotations.deployment),
-          service: mergeKeyValuesWithJson(config.annotations?.service, json.annotations.service),
-          ingress: mergeKeyValuesWithJson(config.annotations?.ingress, json.annotations.ingress),
-        }
-      : null,
-    volumes: json.volumes
-      ? json.volumes.map(volume => {
-          const currentVol = config.volumes?.find(it => it.name === volume.name)
-
-          return {
-            ...volume,
-            id: currentVol?.id ?? uuid(),
-          }
-        })
-      : null,
-  }
-
-  return result
-}
-
-export const mergeJsonWithContainerConfig = (config: ContainerConfig, json: JsonContainerConfig): ContainerConfig => {
-  const concreteConfig: ConcreteContainerConfig = {
-    ...config,
-    secrets: null,
-  }
-
-  const mergedConf = mergeJsonConfigToConcreteContainerConfig(concreteConfig, json)
-  return {
-    ...config,
-    ...mergedConf,
-    secrets: mergeSecretsWithJson(config.secrets, json.secrets),
-  }
 }
 
 export const portToString = (port: ContainerPort): string => {
