+++ web/crux-ui/src/models/index.ts
@@ -8,6 +8,7 @@ export * from './container-config'
 export * from './container-conflict'
 export * from './container-merge'
 export * from './container-errors'
+export * from './json-container'
 export * from './dashboard'
 export * from './deployment'
 export * from './editor'
